Changed object design
This branch is likely to be a mess for a long time.

Now a Block has Attributes, which have Models, which have Textures.
Doing this should help tidy the code, and make it more usable. The final
plan is to be able to support every blockstate file in the game.

At this stage, the only error was in the Add Block form. (It's still
using the old system)