Allow not generating primary models
I realised this as I was typing the guide for v1.1.0

With this, it should be a lot easier to do models that don't naturally
inherit from anything else.

Simply don't type anything into the parent model. It won't create the
non-numbered inheriting model. Then it'll just use the one provided by
the game, and everything else will inherit from that.