Added exponential model (NYI)
Not yet implemented.

This feature might be a little dangerous if you don't know what you're
doing.

It will create a lot of files if you're not careful. A model that
chooses from 6 textures to put on 6 faces will generate over 40,000
model files. Each of those will have an entry into the blockstates file,
and you can increase the size of the blockstates file by 16x if you
wanted with rotation.

Really, what I'm trying to say is, if you don't be careful, then you'll
end up with a lot of files sitting around.
It's called exponential for a reason.

Ctrl-A, Shft-Del.