Major bugfix
Squashed a bug that prevented the last item in a texture group from
getting written.  Kind of a bit hard to explain exactly what was
happening, but be glad that this is now working.

Changed model [name]_0 to be just [name].
Also removed the writing of the template models for now.