Restructure
Oh boy.

Yes, I have taken a bit of a backwards step. I've reintroduced the
one-texture-set model. That is now the base for everything. Above that
we have the ModelCompound. This is a base for objects which contain the
information to create multiple models, but without creating many
objects.

What has been known for the past few commits as "Model" is now
ModelCompoundExponential